add commit

+++ src/main/java/com/community/community/controller/PublishController.java
@@ -1,8 +1,6 @@
 package com.community.community.controller;
 
 import com.community.community.dto.QuestionDTO;
-import com.community.community.mapper.QuestionMapper;
-import com.community.community.mapper.UserMapper;
 import com.community.community.model.Question;
 import com.community.community.model.User;
 import com.community.community.service.QuestionService;
@@ -14,7 +12,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
-import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
 
 @Controller
@@ -25,41 +22,41 @@ public class PublishController {
 
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name="id") Integer id,
-                       Model model){
+    public String edit(@PathVariable(name = "id") Long id,
+                       Model model) {
         QuestionDTO question = questionService.getById(id);
-        model.addAttribute("title",question.getTitle());
-        model.addAttribute("description",question.getDescription());
-        model.addAttribute("tag",question.getTag());
-        model.addAttribute("id",question.getId());
+        model.addAttribute("title", question.getTitle());
+        model.addAttribute("description", question.getDescription());
+        model.addAttribute("tag", question.getTag());
+        model.addAttribute("id", question.getId());
         return "publish";
     }
 
     @GetMapping("/publish")
-    public String publish(){
+    public String publish() {
         return "publish";
     }
 
     @PostMapping("/publish")
-    public String doPublish(@RequestParam(value = "title",required = false) String title,
-                            @RequestParam(value = "description",required = false) String description,
-                            @RequestParam(value = "tag",required = false) String tag,
-                            @RequestParam(value = "id",required = false) Integer id,
+    public String doPublish(@RequestParam(value = "title", required = false) String title,
+                            @RequestParam(value = "description", required = false) String description,
+                            @RequestParam(value = "tag", required = false) String tag,
+                            @RequestParam(value = "id", required = false) Long id,
                             HttpServletRequest request,
-                            Model model){
-        model.addAttribute("title",title);
-        model.addAttribute("description",description);
-        model.addAttribute("tag",tag);
-        if(title==null||title==""){
-            model.addAttribute("error","标题不能为空");
+                            Model model) {
+        model.addAttribute("title", title);
+        model.addAttribute("description", description);
+        model.addAttribute("tag", tag);
+        if (title == null || title == "") {
+            model.addAttribute("error", "标题不能为空");
             return "/publish";
         }
-        if(description==null||description==""){
-            model.addAttribute("error","描述不能为空");
+        if (description == null || description == "") {
+            model.addAttribute("error", "描述不能为空");
             return "/publish";
         }
-        if(tag==null||tag==""){
-            model.addAttribute("error","标签不能为空");
+        if (tag == null || tag == "") {
+            model.addAttribute("error", "标签不能为空");
             return "/publish";
         }
 
@@ -69,14 +66,14 @@ public class PublishController {
         question.setTag(tag);
         //user
         User user = (User) request.getSession().getAttribute("user");
-        if(user==null){
-            model.addAttribute("error","用户未登陆");
+        if (user == null) {
+            model.addAttribute("error", "用户未登陆");
             return "/publish";
         }
-        if(id==null) {
+        if (id == null) {
             question.setCreator(user.getId());
         }
-        questionService.createOrUpdate(question,id);
+        questionService.createOrUpdate(question, id);
         return "redirect:/";
     }
 }
